ModLog
Modlogs oversight & added modlogs to automod

+++ functions/performAutomod.ts
@@ -1,4 +1,4 @@
-import { Client, Message, MessageEmbed, Permissions } from "discord.js";
+import { Client, Message, MessageEmbed, Permissions, TextChannel } from "discord.js";
 import automodConfig from "../models/automodConfig";
 import Guild from "../models/guild";
 import Config from "../models/config";
@@ -64,8 +64,35 @@ export = async function performAutomod(message: Message, client: Client) {
                         date: Date.now(),
                     })
                     newCases.save()
+                    let messageA
+                    if (message.content.length < 450) {
+                        messageA = message.content
+                    } else {
+                        messageA = "Too long!"
+                    }
+                    const modLogEmbed = new MessageEmbed()
+                        .setAuthor({ name: `Member Muted - ${message.author.tag}`, iconURL: message.author.displayAvatarURL({ dynamic: true }) || "" })
+                        .setThumbnail(message.author.displayAvatarURL({ dynamic: true }) || "")
+                        .setColor(configuration.color)
+                        .setTimestamp()
+                        .setDescription(`<:user:977391493218181120> **User:** ${message.author.tag}
+                        > [${message.author.id}]
+                        > [<@${message.author.id}>]
+                        <:folder:977391492790362173> **Mod:** \`AUTO-MOD\`
+                        <:pencil:977391492916207636> **Action:** Mute
+                        > [**Case:** #${caseNumberSet}]
+                        > [**Duration:** 10 Minutes]
+                        > [**Message:** \`${messageA}\`]
+                        **Reason:** Sending unauthorized links. Automatic mute due to excessive warnings.
+                        **Channel:** <#${message.channel?.id}>
+                        **Date:** <t:${Math.round(Date.now() / 1000)}:D>`)
+                    const channel = message.guild?.channels.cache.find((c: any) => c.id === configuration.modLogChannel);
+                    if (!channel) { return; }
+                    if(message.guild?.me?.permissionsIn(channel).has(Permissions.FLAGS.SEND_MESSAGES)) { 
+                        (message.guild?.channels.cache.find((c: any) => c.id === channel?.id) as TextChannel).send({ embeds: [modLogEmbed] })
+                     }
                     const warnEmbed = new MessageEmbed()
-                        .setDescription(`**Reason:** Sending a link to an unauthorized website.`)
+                        .setDescription(`**Reason:** Sending unauthorized links.`)
                         .setColor(configuration.embedColor)
                     return message.reply({ content: "<:arrow_right:967329549912248341> You have been automatically muted.", embeds: [warnEmbed] }).then(msg => {
                         setTimeout(() => {
@@ -88,6 +115,32 @@ export = async function performAutomod(message: Message, client: Client) {
                     date: Date.now(),
                 })
                 newCases.save()
+                let messageA
+                if (message.content.length < 450) {
+                    messageA = message.content
+                } else {
+                    messageA = "Too long!"
+                }
+                const modLogEmbed = new MessageEmbed()
+                    .setAuthor({ name: `Member Warned - ${message.author.tag}`, iconURL: message.author.displayAvatarURL({ dynamic: true }) || "" })
+                    .setThumbnail(message.author.displayAvatarURL({ dynamic: true }) || "")
+                    .setColor(configuration.color)
+                    .setTimestamp()
+                    .setDescription(`<:user:977391493218181120> **User:** ${message.author.tag}
+                    > [${message.author.id}]
+                    > [<@${message.author.id}>]
+                    <:folder:977391492790362173> **Mod:** \`AUTO-MOD\`
+                    <:pencil:977391492916207636> **Action:** Warn
+                    > [**Case:** #${caseNumberSet}]
+                    > [**Message:** \`${messageA}\`]
+                    **Reason:** Sending unauthorized links.
+                    **Channel:** <#${message.channel?.id}>
+                    **Date:** <t:${Math.round(Date.now() / 1000)}:D>`)
+                const channel = message.guild?.channels.cache.find((c: any) => c.id === configuration.modLogChannel);
+                if (!channel) { return; }
+                if(message.guild?.me?.permissionsIn(channel).has(Permissions.FLAGS.SEND_MESSAGES)) { 
+                    (message.guild?.channels.cache.find((c: any) => c.id === channel?.id) as TextChannel).send({ embeds: [modLogEmbed] })
+                 }
                 const warnEmbed = new MessageEmbed()
                     .setDescription(`**Reason:** Sending a link to an unauthorized website.`)
                     .setColor(configuration.embedColor)
@@ -104,16 +157,16 @@ export = async function performAutomod(message: Message, client: Client) {
     }
 
     //Scam Blocker
-    if(aConfig.blockScams === true) {
+    if (aConfig.blockScams === true) {
         let isScammer = false
-        if(!message.guild?.me?.permissions.has(Permissions.FLAGS.BAN_MEMBERS)) { return; }
-        for(const scams of badlinks) {
-            if(message.content.toLowerCase().includes(scams.toLowerCase())) {
+        if (!message.guild?.me?.permissions.has(Permissions.FLAGS.BAN_MEMBERS)) { return; }
+        for (const scams of badlinks) {
+            if (message.content.toLowerCase().includes(scams.toLowerCase())) {
                 isScammer = true
                 break;
             }
         }
-        if(isScammer == true) {
+        if (isScammer == true) {
             const newCases = await new Cases({
                 guildID: message.guild?.id,
                 userID: message.author.id,
@@ -125,20 +178,47 @@ export = async function performAutomod(message: Message, client: Client) {
                 date: Date.now(),
             })
             newCases.save()
+            let messageA
+            if (message.content.length < 450) {
+                messageA = message.content
+            } else {
+                messageA = "Too long!"
+            }
+            const modLogEmbed = new MessageEmbed()
+                .setAuthor({ name: `Member Banned - ${message.author.tag}`, iconURL: message.author.displayAvatarURL({ dynamic: true }) || "" })
+                .setThumbnail(message.author.displayAvatarURL({ dynamic: true }) || "")
+                .setColor(configuration.color)
+                .setTimestamp()
+                .setDescription(`<:user:977391493218181120> **User:** ${message.author.tag}
+                > [${message.author.id}]
+                > [<@${message.author.id}>]
+                <:folder:977391492790362173> **Mod:** \`AUTO-MOD\`
+                <:pencil:977391492916207636> **Action:** Ban
+                > [**Case:** #${caseNumberSet}]
+                > [**Duration:** Permanent]
+                > [**Message:** \`${messageA}\`]
+                **Reason:** Sending scam links.
+                **Channel:** <#${message.channel?.id}>
+                **Date:** <t:${Math.round(Date.now() / 1000)}:D>`)
+            const channel = message.guild?.channels.cache.find((c: any) => c.id === configuration.modLogChannel);
+            if (!channel) { return; }
+            if(message.guild?.me?.permissionsIn(channel).has(Permissions.FLAGS.SEND_MESSAGES)) { 
+                (message.guild?.channels.cache.find((c: any) => c.id === channel?.id) as TextChannel).send({ embeds: [modLogEmbed] })
+             }
             const warnEmbed = new MessageEmbed()
-                    .setDescription(`**Reason:** Sending scan links.`)
-                    .setColor(configuration.embedColor)
-                return message.channel.send({ content: `<:arrow_right:967329549912248341> ${message.author.tag} has been banned.`, embeds: [warnEmbed] }).then(msg => {
-                    setTimeout(() => {
-                        if (msg.deletable) {
-                            msg.delete()
-                        }
-                    }, 10000)
-                    message.delete().catch((err: Error) => console.error(err));
-                    if(message.member?.bannable) {
-                        message.member?.ban({ reason: "Sending scam links", days: 7 })
+                .setDescription(`**Reason:** Sending scan links.`)
+                .setColor(configuration.embedColor)
+            return message.channel.send({ content: `<:arrow_right:967329549912248341> ${message.author.tag} has been banned.`, embeds: [warnEmbed] }).then(msg => {
+                setTimeout(() => {
+                    if (msg.deletable) {
+                        msg.delete()
                     }
-                })
+                }, 10000)
+                message.delete().catch((err: Error) => console.error(err));
+                if (message.member?.bannable) {
+                    message.member?.ban({ reason: "Sending scam links", days: 7 })
+                }
+            })
         }
     }
 
@@ -150,8 +230,6 @@ export = async function performAutomod(message: Message, client: Client) {
                 if (warns !== 0) {
                     remainder = warns % configuration.warnsBeforeMute
                 }
-                console.log(warns)
-                console.log(remainder)
                 if (remainder == 0) {
                     const newCases = await new Cases({
                         guildID: message.guild?.id,
@@ -164,6 +242,33 @@ export = async function performAutomod(message: Message, client: Client) {
                         date: Date.now(),
                     })
                     newCases.save()
+                    let messageA
+                    if (message.content.length < 450) {
+                        messageA = message.content
+                    } else {
+                        messageA = "Too long!"
+                    }
+                    const modLogEmbed = new MessageEmbed()
+                        .setAuthor({ name: `Member Muted - ${message.author.tag}`, iconURL: message.author.displayAvatarURL({ dynamic: true }) || "" })
+                        .setThumbnail(message.author.displayAvatarURL({ dynamic: true }) || "")
+                        .setColor(configuration.color)
+                        .setTimestamp()
+                        .setDescription(`<:user:977391493218181120> **User:** ${message.author.tag}
+                        > [${message.author.id}]
+                        > [<@${message.author.id}>]
+                        <:folder:977391492790362173> **Mod:** \`AUTO-MOD\`
+                        <:pencil:977391492916207636> **Action:** Mute
+                        > [**Case:** #${caseNumberSet}]
+                        > [**Message:** \`${messageA}\`]
+                        **Reason:** Exceeding max mentions. Automatic mute due to excessive warnings.
+                        **Channel:** <#${message.channel?.id}>
+                        **Date:** <t:${Math.round(Date.now() / 1000)}:D>`)
+                    const channel = message.guild?.channels.cache.find((c: any) => c.id === configuration.modLogChannel);
+                    if (!channel) { return; }
+                    if(message.guild?.me?.permissionsIn(channel).has(Permissions.FLAGS.SEND_MESSAGES)) { 
+                        (message.guild?.channels.cache.find((c: any) => c.id === channel?.id) as TextChannel).send({ embeds: [modLogEmbed] })
+                     }
+    
                     const warnEmbed = new MessageEmbed()
                         .setDescription(`**Reason:** Exceeding max mentions.`)
                         .setColor(configuration.embedColor)
@@ -188,6 +293,32 @@ export = async function performAutomod(message: Message, client: Client) {
                     date: Date.now(),
                 })
                 newCases.save()
+                let messageA
+                if (message.content.length < 450) {
+                    messageA = message.content
+                } else {
+                    messageA = "Too long!"
+                }
+                const modLogEmbed = new MessageEmbed()
+                    .setAuthor({ name: `Member Warned - ${message.author.tag}`, iconURL: message.author.displayAvatarURL({ dynamic: true }) || "" })
+                    .setThumbnail(message.author.displayAvatarURL({ dynamic: true }) || "")
+                    .setColor(configuration.color)
+                    .setTimestamp()
+                    .setDescription(`<:user:977391493218181120> **User:** ${message.author.tag}
+                    > [${message.author.id}]
+                    > [<@${message.author.id}>]
+                    <:folder:977391492790362173> **Mod:** \`AUTO-MOD\`
+                    <:pencil:977391492916207636> **Action:** Warn
+                    > [**Case:** #${caseNumberSet}]
+                    > [**Message:** \`${messageA}\`]
+                    **Reason:** Exceeding max mentions.
+                    **Channel:** <#${message.channel?.id}>
+                    **Date:** <t:${Math.round(Date.now() / 1000)}:D>`)
+                const channel = message.guild?.channels.cache.find((c: any) => c.id === configuration.modLogChannel);
+                if (!channel) { return; }
+                if(message.guild?.me?.permissionsIn(channel).has(Permissions.FLAGS.SEND_MESSAGES)) { 
+                    (message.guild?.channels.cache.find((c: any) => c.id === channel?.id) as TextChannel).send({ embeds: [modLogEmbed] })
+                }
                 const warnEmbed = new MessageEmbed()
                     .setDescription(`**Reason:** Exceeding max mentions.`)
                     .setColor(configuration.embedColor)
